+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unpublished
+### Added
+- callback to be called on cancellation can now be provided via start()
+  second argument argument
+- Cancel calls can recieve a reason argument, passed down to cancel callback or
+  used in abortion promise
 ### Changed
+- non-started Timeouts can be rejected now
 - Timeout cancellation will result in Promise cancellation with reject error.
 
 ## [2.1.0] - 2022-01-24